backend/requirements.txt
google-generativeai
python-dotenv
Pillow
requests
openai>=1.0.0
